+++ components/src/main/java/com/mercadolibre/android/andesui/thumbnail/shape/AndesThumbnailShapeInterface.kt
@@ -1,0 +1,16 @@
+package com.mercadolibre.android.andesui.thumbnail.shape
+
+/**
+ * defines all the possible shapes of a Thumbnail.
+ */
+interface AndesThumbnailShapeInterface
+
+/**
+ * This class represent the circle Thumbnail shape.
+ */
+internal class AndesCircleShapeThumbnail : AndesThumbnailShapeInterface
+
+/**
+ * This class represent the square Thumbnail shape.
+ */
+internal class AndesSquareShapeThumbnail : AndesThumbnailShapeInterface