+++ components/src/main/java/com/mercadolibre/android/andesui/thumbnail/size/AndesThumbnailSizeInterface.kt
@@ -4,13 +4,16 @@ import android.content.Context
 import com.mercadolibre.android.andesui.R
 
 /**
- * Defines all size related properties that an [AndesThumbnail] needs to be drawn properly.
+ * Defines all size related properties that an AndesThumbnail needs to be drawn properly.
  * Those properties change depending on the size of the thumbnail.
  */
 internal interface AndesThumbnailSizeInterface {
     fun diameter(context: Context): Float
     fun iconSize(context: Context): Float
     fun radiusSize(context: Context): Float
+    fun textSize(context: Context): Float
+    fun textHeight(context: Context): Float
+    fun textWidth(context: Context): Float
 }
 
 /**
@@ -23,6 +26,12 @@ internal class Andes24ThumbnailSize : AndesThumbnailSizeInterface {
         context.resources.getDimension(R.dimen.andes_thumbnail_icon_size_16)
     override fun radiusSize(context: Context): Float =
         context.resources.getDimension(R.dimen.andes_thumbnail_corner_radius_3)
+    override fun textSize(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_11)
+    override fun textHeight(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_11)
+    override fun textWidth(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_width_20)
 }
 
 /**
@@ -35,6 +44,12 @@ internal class Andes32ThumbnailSize : AndesThumbnailSizeInterface {
         context.resources.getDimension(R.dimen.andes_thumbnail_icon_size_20)
     override fun radiusSize(context: Context): Float =
         context.resources.getDimension(R.dimen.andes_thumbnail_corner_radius_3)
+    override fun textSize(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_14)
+    override fun textHeight(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_14)
+    override fun textWidth(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_width_28)
 }
 
 /**
@@ -47,6 +62,11 @@ internal class Andes40ThumbnailSize : AndesThumbnailSizeInterface {
         context.resources.getDimension(R.dimen.andes_thumbnail_icon_size_24)
     override fun radiusSize(context: Context): Float =
         context.resources.getDimension(R.dimen.andes_thumbnail_corner_radius_4)
+    override fun textSize(context: Context): Float = context.resources.getDimension(R.dimen.andes_thumbnail_text_size_16)
+    override fun textHeight(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_16)
+    override fun textWidth(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_width_32)
 }
 
 /**
@@ -59,6 +79,11 @@ internal class Andes48ThumbnailSize : AndesThumbnailSizeInterface {
         context.resources.getDimension(R.dimen.andes_thumbnail_icon_size_24)
     override fun radiusSize(context: Context): Float =
         context.resources.getDimension(R.dimen.andes_thumbnail_corner_radius_4)
+    override fun textSize(context: Context): Float = context.resources.getDimension(R.dimen.andes_thumbnail_text_size_20)
+    override fun textHeight(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_20)
+    override fun textWidth(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_width_36)
 }
 
 /**
@@ -71,6 +96,11 @@ internal class Andes56ThumbnailSize : AndesThumbnailSizeInterface {
         context.resources.getDimension(R.dimen.andes_thumbnail_icon_size_32)
     override fun radiusSize(context: Context): Float =
         context.resources.getDimension(R.dimen.andes_thumbnail_corner_radius_4)
+    override fun textSize(context: Context): Float = context.resources.getDimension(R.dimen.andes_thumbnail_text_size_24)
+    override fun textHeight(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_24)
+    override fun textWidth(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_width_44)
 }
 
 /**
@@ -83,6 +113,11 @@ internal class Andes64ThumbnailSize : AndesThumbnailSizeInterface {
         context.resources.getDimension(R.dimen.andes_thumbnail_icon_size_32)
     override fun radiusSize(context: Context): Float =
         context.resources.getDimension(R.dimen.andes_thumbnail_corner_radius_4)
+    override fun textSize(context: Context): Float = context.resources.getDimension(R.dimen.andes_thumbnail_text_size_28)
+    override fun textHeight(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_28)
+    override fun textWidth(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_width_52)
 }
 
 /**
@@ -96,6 +131,11 @@ internal class Andes72ThumbnailSize : AndesThumbnailSizeInterface {
         context.resources.getDimension(R.dimen.andes_thumbnail_icon_size_32)
     override fun radiusSize(context: Context): Float =
         context.resources.getDimension(R.dimen.andes_thumbnail_corner_radius_5)
+    override fun textSize(context: Context): Float = context.resources.getDimension(R.dimen.andes_thumbnail_text_size_32)
+    override fun textHeight(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_32)
+    override fun textWidth(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_width_60)
 }
 
 /**
@@ -109,4 +149,9 @@ internal class Andes80ThumbnailSize : AndesThumbnailSizeInterface {
         context.resources.getDimension(R.dimen.andes_thumbnail_icon_size_48)
     override fun radiusSize(context: Context): Float =
         context.resources.getDimension(R.dimen.andes_thumbnail_corner_radius_5)
+    override fun textSize(context: Context): Float = context.resources.getDimension(R.dimen.andes_thumbnail_text_size_36)
+    override fun textHeight(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_size_36)
+    override fun textWidth(context: Context): Float =
+        context.resources.getDimension(R.dimen.andes_thumbnail_text_width_64)
 }