+++ components/src/test/java/com/mercadolibre/android/andesui/thumbnail/AndesThumbnailBadgeConfigurationFactoryTest.kt
@@ -2,6 +2,7 @@ package com.mercadolibre.android.andesui.thumbnail
 
 import android.content.Context
 import android.content.res.ColorStateList
+import android.view.View
 import androidx.core.content.ContextCompat
 import androidx.test.core.app.ApplicationProvider
 import com.mercadolibre.android.andesui.R
@@ -12,6 +13,7 @@ import com.mercadolibre.android.andesui.badge.AndesBadgePill
 import com.mercadolibre.android.andesui.badge.icontype.AndesBadgeIconType
 import com.mercadolibre.android.andesui.badge.size.AndesBadgePillSize
 import com.mercadolibre.android.andesui.badge.type.AndesBadgeType
+import com.mercadolibre.android.andesui.thumbnail.assetType.AndesThumbnailAssetType
 import com.mercadolibre.android.andesui.thumbnail.badge.component.AndesDotThumbnailBadgeComponent
 import com.mercadolibre.android.andesui.thumbnail.badge.component.AndesIconPillThumbnailBadgeComponent
 import com.mercadolibre.android.andesui.thumbnail.badge.component.AndesPillThumbnailBadgeComponent
@@ -22,8 +24,8 @@ import com.mercadolibre.android.andesui.thumbnail.badge.size.AndesThumbnailBadge
 import com.mercadolibre.android.andesui.thumbnail.badge.size.AndesThumbnailBadgePillSize
 import com.mercadolibre.android.andesui.thumbnail.badge.type.AndesThumbnailBadgeType
 import com.mercadolibre.android.andesui.thumbnail.size.AndesThumbnailSize
-import com.mercadolibre.android.andesui.thumbnail.type.AndesThumbnailType
 import com.mercadolibre.android.andesui.utils.Constants.TEST_ANDROID_VERSION_CODE
+import com.mercadolibre.android.andesui.utils.assertNotEquals
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -57,8 +59,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
                 false,
                 AndesThumbnailBadgePillSize.SIZE_64
             ),
-            thumbnailType = expectedType
-        )
+            thumbnailType = expectedType,
+        "")
 
         // WHEN
         val config = AndesThumbnailBadgeConfigurationFactory.create(context, attrs)
@@ -68,9 +70,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
             val expectedColor = expectedBadgeColor.iconType.type.primaryColor().colorInt(context)
             badgeColor assertEquals expectedColor
             thumbnailSize assertEquals AndesThumbnailSize.SIZE_64
-            thumbnailType assertEquals AndesThumbnailType.ICON
+            assetType assertEquals AndesThumbnailAssetType.Icon(expectedImage)
             thumbnailTintColor assertEquals ColorStateList.valueOf(expectedColor)
-            image assertEquals expectedImage
             badgeOutline assertEquals context.resources.getDimensionPixelSize(R.dimen.andes_thumbnail_badge_outline_3)
             assert(badgeComponent is AndesPillThumbnailBadgeComponent)
             with(badgeComponent.getView(context) as AndesBadgePill) {
@@ -97,7 +98,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
                 false,
                 AndesThumbnailBadgePillSize.SIZE_56
             ),
-            thumbnailType = expectedType
+            thumbnailType = expectedType,
+            ""
         )
 
         // WHEN
@@ -108,9 +110,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
             val expectedColor = expectedBadgeColor.iconType.type.primaryColor().colorInt(context)
             badgeColor assertEquals expectedColor
             thumbnailSize assertEquals AndesThumbnailSize.SIZE_56
-            thumbnailType assertEquals AndesThumbnailType.IMAGE_CIRCLE
+            assetType assertEquals AndesThumbnailAssetType.Image(expectedImage)
             thumbnailTintColor assertEquals null
-            image assertEquals expectedImage
             badgeOutline assertEquals context.resources.getDimensionPixelSize(R.dimen.andes_thumbnail_badge_outline_2)
             assert(badgeComponent is AndesPillThumbnailBadgeComponent)
             with(badgeComponent.getView(context) as AndesBadgePill) {
@@ -134,7 +135,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
                 expectedBadgeColor,
                 AndesThumbnailBadgeDotSize.SIZE_24
             ),
-            thumbnailType = expectedType
+            thumbnailType = expectedType,
+            ""
         )
 
         // WHEN
@@ -145,9 +147,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
             val expectedColor = expectedBadgeColor.iconType.type.primaryColor().colorInt(context)
             badgeColor assertEquals expectedColor
             thumbnailSize assertEquals AndesThumbnailSize.SIZE_24
-            thumbnailType assertEquals AndesThumbnailType.ICON
+            assetType assertEquals AndesThumbnailAssetType.Icon(expectedImage)
             thumbnailTintColor assertEquals ColorStateList.valueOf(expectedColor)
-            image assertEquals expectedImage
             badgeOutline assertEquals context.resources.getDimensionPixelSize(R.dimen.andes_thumbnail_badge_outline_2)
             assert(badgeComponent is AndesDotThumbnailBadgeComponent)
             with(badgeComponent.getView(context) as AndesBadgeDot) {
@@ -168,7 +169,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
                 expectedBadgeColor,
                 AndesThumbnailBadgeDotSize.SIZE_32
             ),
-            thumbnailType = expectedType
+            thumbnailType = expectedType,
+            ""
         )
 
         // WHEN
@@ -179,9 +181,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
             val expectedColor = expectedBadgeColor.iconType.type.primaryColor().colorInt(context)
             badgeColor assertEquals expectedColor
             thumbnailSize assertEquals AndesThumbnailSize.SIZE_32
-            thumbnailType assertEquals AndesThumbnailType.IMAGE_CIRCLE
+            assetType assertEquals AndesThumbnailAssetType.Image(expectedImage)
             thumbnailTintColor assertEquals null
-            image assertEquals expectedImage
             badgeOutline assertEquals context.resources.getDimensionPixelSize(R.dimen.andes_thumbnail_badge_outline_2)
             assert(badgeComponent is AndesDotThumbnailBadgeComponent)
             with(badgeComponent.getView(context) as AndesBadgeDot) {
@@ -202,7 +203,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
                 expectedBadgeColor,
                 AndesThumbnailBadgePillSize.SIZE_72
             ),
-            thumbnailType = expectedType
+            thumbnailType = expectedType,
+            ""
         )
 
         // WHEN
@@ -213,9 +215,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
             val expectedColor = expectedBadgeColor.iconType.type.primaryColor().colorInt(context)
             badgeColor assertEquals expectedColor
             thumbnailSize assertEquals AndesThumbnailSize.SIZE_72
-            thumbnailType assertEquals AndesThumbnailType.IMAGE_CIRCLE
+            assetType assertEquals AndesThumbnailAssetType.Image(expectedImage)
             thumbnailTintColor assertEquals null
-            image assertEquals expectedImage
             badgeOutline assertEquals context.resources.getDimensionPixelSize(R.dimen.andes_thumbnail_badge_outline_3)
             assert(badgeComponent is AndesIconPillThumbnailBadgeComponent)
             with(badgeComponent.getView(context) as AndesBadgeIconPill) {
@@ -237,7 +238,8 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
                 expectedBadgeColor,
                 AndesThumbnailBadgePillSize.SIZE_48
             ),
-            thumbnailType = expectedType
+            thumbnailType = expectedType,
+            ""
         )
 
         // WHEN
@@ -248,13 +250,117 @@ class AndesThumbnailBadgeConfigurationFactoryTest {
             val expectedColor = expectedBadgeColor.iconType.type.primaryColor().colorInt(context)
             badgeColor assertEquals expectedColor
             thumbnailSize assertEquals AndesThumbnailSize.SIZE_48
-            thumbnailType assertEquals AndesThumbnailType.ICON
+            assetType assertEquals AndesThumbnailAssetType.Icon(expectedImage)
             thumbnailTintColor assertEquals ColorStateList.valueOf(expectedColor)
-            image assertEquals expectedImage
             badgeOutline assertEquals context.resources.getDimensionPixelSize(R.dimen.andes_thumbnail_badge_outline_2)
+            badgeVisibility assertEquals View.VISIBLE
             assert(badgeComponent is AndesIconPillThumbnailBadgeComponent)
             with(badgeComponent.getView(context) as AndesBadgeIconPill) {
                 type assertEquals AndesBadgeIconType.WARNING
+                size assertEquals AndesBadgePillSize.SMALL
+            }
+        }
+    }
+
+    @Test
+    fun `Config Icon Pill, Text, SIZE_56, WARNING`() {
+        // GIVEN
+        val expectedBadgeColor = AndesBadgeIconType.WARNING
+        val expectedType = AndesThumbnailBadgeType.Text
+        attrs = AndesThumbnailBadgeAttrs(
+            image = null,
+            badge = AndesThumbnailBadgeComponent.IconPill(
+                expectedBadgeColor,
+                AndesThumbnailBadgePillSize.SIZE_56
+            ),
+            thumbnailType = expectedType,
+            "AB"
+        )
+
+        // WHEN
+        val config = AndesThumbnailBadgeConfigurationFactory.create(context, attrs)
+
+        // THEN
+        with(config) {
+            val expectedColor = expectedBadgeColor.iconType.type.primaryColor().colorInt(context)
+            badgeColor assertEquals expectedColor
+            thumbnailSize assertEquals AndesThumbnailSize.SIZE_56
+            assetType assertEquals AndesThumbnailAssetType.Text("AB")
+            thumbnailTintColor assertEquals ColorStateList.valueOf(expectedColor)
+            badgeOutline assertEquals context.resources.getDimensionPixelSize(R.dimen.andes_thumbnail_badge_outline_2)
+            badgeVisibility assertNotEquals View.GONE
+            assert(badgeComponent is AndesIconPillThumbnailBadgeComponent)
+            with(badgeComponent.getView(context) as AndesBadgeIconPill) {
+                type assertEquals AndesBadgeIconType.WARNING
+                size assertEquals AndesBadgePillSize.SMALL
+            }
+        }
+    }
+
+    @Test
+    fun `Config Icon Badge, Text, SIZE_80, SUCCESS`() {
+        // GIVEN
+        val expectedBadgeColor = AndesBadgeIconType.SUCCESS
+        val expectedType = AndesThumbnailBadgeType.Text
+        attrs = AndesThumbnailBadgeAttrs(
+            image = null,
+            badge = AndesThumbnailBadgeComponent.Pill(
+                expectedBadgeColor,
+                "CD",
+                true,
+                AndesThumbnailBadgePillSize.SIZE_80
+            ),
+            thumbnailType = expectedType,
+            "AB"
+        )
+
+        // WHEN
+        val config = AndesThumbnailBadgeConfigurationFactory.create(context, attrs)
+
+        // THEN
+        with(config) {
+            val expectedColor = expectedBadgeColor.iconType.type.primaryColor().colorInt(context)
+            badgeColor assertEquals expectedColor
+            thumbnailSize assertEquals AndesThumbnailSize.SIZE_80
+            assetType assertEquals AndesThumbnailAssetType.Text("AB")
+            thumbnailTintColor assertEquals ColorStateList.valueOf(expectedColor)
+            badgeOutline assertEquals context.resources.getDimensionPixelSize(R.dimen.andes_thumbnail_badge_outline_3)
+            badgeVisibility assertEquals View.VISIBLE
+            assert(badgeComponent is AndesPillThumbnailBadgeComponent)
+            with(badgeComponent.getView(context) as AndesBadgePill) {
+                type assertEquals AndesBadgeType.SUCCESS
+            }
+        }
+    }
+
+    @Test
+    fun `Config Icon Pill, Text, SIZE_56, SUCCESS`() {
+        // GIVEN
+        val expectedBadgeColor = AndesBadgeIconType.SUCCESS
+        val expectedType = AndesThumbnailBadgeType.Text
+        attrs = AndesThumbnailBadgeAttrs(
+            image=null,
+            badge = AndesThumbnailBadgeComponent.IconPill(
+                expectedBadgeColor,
+                AndesThumbnailBadgePillSize.SIZE_56
+            ),
+            thumbnailType = expectedType,
+            "AB"
+        )
+
+        // WHEN
+        val config = AndesThumbnailBadgeConfigurationFactory.create(context, attrs)
+
+        // THEN
+        with(config) {
+            val expectedColor = expectedBadgeColor.iconType.type.primaryColor().colorInt(context)
+            badgeColor assertEquals expectedColor
+            thumbnailSize assertEquals AndesThumbnailSize.SIZE_56
+            assetType assertEquals AndesThumbnailAssetType.Text(text)
+            badgeOutline assertEquals context.resources.getDimensionPixelSize(R.dimen.andes_thumbnail_badge_outline_2)
+            assert(badgeComponent is AndesIconPillThumbnailBadgeComponent)
+            with(badgeComponent.getView(context) as AndesBadgeIconPill) {
+                type assertEquals AndesBadgeIconType.SUCCESS
                 size assertEquals AndesBadgePillSize.SMALL
             }
         }