+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v6.1.0
+## 🛠 Fixes
+- Fix crash in demoapp caused by an activity without the `exported` property 
+- Fix Overlap between tooltip title and dismiss button when title text is large | Author: [@meligeo](https://github.com/meligeo)
+## 🚀 Feature
+- Now the `AndesThumbnail` and `AndesThumbnailBadge` as a new text type that allow to create a Thumbnail with letters instead of an image | Author: [@MLcaro](https://github.com/MLcaro)
+- Added new attributes to  `AndesThumbnail` receives the shape and the asset type separately| Author: [@MLcaro](https://github.com/MLcaro)
+
+# v6.0.0
+## Migrated
+- Migrated to target API 32 (Android 12)
+
 # v5.13.2
 ## 🛠 Fixes
 - Fix AndesTextView with AndesMoneyAmount semibold | Author: [@eportasML](https://github.com/eportasML)